Подключил liquibase

+++ order/src/main/resources/db/005_ddl_create_card_table.sql
@@ -1,6 +1,6 @@
 create table card (
     id serial primary key not null,
     number varchar unique not null,
-    discount double not null,
+    discount double precision not null,
     customer_id int not null references customer(id)
 );
